requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.6.0
flask-cors==4.0.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0
dateparser==1.2.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
xhtml2pdf==0.2.15
openpyxl==3.1.2
requests==2.31.0
google-genai==0.3.0
python-dotenv==1.0.1
